optimize load strategy
